Implementing Taum And Bday - changing to long and improving tests

+++ puzzles-java/src/main/java/pedrotoliveira/puzzles/hackerrank/algorithm/implementation/Gifts.java
@@ -6,19 +6,19 @@ package pedrotoliveira.puzzles.hackerrank.algorithm.implementation;
  */
 public class Gifts {
 
-	private final int quantity;
-	private final int cost;
+	private final long quantity;
+	private final long cost;
 
-	public Gifts(int quantity, int cost) {
+	public Gifts(long quantity, long cost) {
 		this.quantity = quantity;
 		this.cost = cost;
 	}
 
-	public int getQuantity() {
+	public long getQuantity() {
 		return quantity;
 	}
 
-	public int getCost() {
+	public long getCost() {
 		return cost;
 	}
